==> notebooks/01_preprocess.ipynb <==
DATA_PATH = '../data/raw/stack-overflow-developer-survey-2021/survey_results_public.csv'
EXPORT_PATH = "../data/processed/01_preprocessed_df.pkl"

REPLACE_DICT = {
    'YearsCodePro': {'Less than 1 year': 0, 'More than 50 years': 51},
    'YearsCode':    {'Less than 1 year': 0, 'More than 50 years': 51}}

import numpy as np
import pandas as pd

from scripts.preprocessing import split_answers

## Preprocess Data

# Read data
raw_df = pd.read_csv(DATA_PATH)
df = raw_df.copy()

### 1. Replace string values and parse
for col, replacement in REPLACE_DICT.items():
    df[col] = df[col].replace(replacement).astype(np.float32)

### 2. Split multiple answers
object_cols = df.select_dtypes(include='object').columns.tolist()
for col in object_cols:
    df[col] = split_answers(df[col])

### Export

df.to_pickle(EXPORT_PATH)
